Refactor project structure, Add requirements.txt
