+++ .gitignore
@@ -1,2 +1,4 @@
 __pycache__
 *.svg
+env
+.idea
